Removed swagger plugin
Swagger is really geared towards projects with an API that expose a set of well
defined objects and set of operation on those objects.

Between our Record/FormClass type system and CUID generation, it seems unlikely
that we will be able to autogenerate client libraries and docs as I had hoped.