Updated AuthorizationTest to reflect change in versioning strategy
A transactional update to a workspace may update several resources, and each will receive the
same version number.

In this test, for example, a workspace is created together with an ACR. If a user where to sync to version one
and only receive the workspace, but not the ACR, their client would be left in an inconsistent state.