AI-751 Split up SchemaDTO into ActivityDTO and ActivityFormDTO
This has become necessary as the number and complexity of forms visible
to users is growing quite rapidly and sending all forms and their indicators, attributes, etc
over the wire is degrading performance across the platform.

Most importantly, this is needed to support the MHFS use case of a form with 700+ form fields.
This works when modeled as a form class, but if we have to query all the attributes/indicators,
the server mostly times out.