Removed test that no longer works because of the inclusion of UI code in presenter
(My own experience with the Presenter/View pattern is that it's really not worth the
effort unless the logic is super, super complicated. Otherwise you end up producing
three times the code to maintain the split and the result is hard to understand, and
you are still only able to test the most basic elements with unit tests.

I think the reality is that the distinction between UI "logic" and "view" is artificial
in most cases and you are better of just building a single highly cohesive component.